+++ modules/camera_iphone/camera_ios.h
