+++ modules/gamecenter/game_center_delegate.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  in_app_store.h                                                       */
+/*  game_center_delegate.h                                               */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,54 +28,8 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifdef STOREKIT_ENABLED
+#import <GameKit/GameKit.h>
 
-#ifndef IN_APP_STORE_H
-#define IN_APP_STORE_H
+@interface GodotGameCenterDelegate : NSObject <GKGameCenterControllerDelegate>
 
-#include "core/object/class_db.h"
-
-#ifdef __OBJC__
-@class GodotProductsDelegate;
-@class GodotTransactionsObserver;
-
-typedef GodotProductsDelegate InAppStoreProductDelegate;
-typedef GodotTransactionsObserver InAppStoreTransactionObserver;
-#else
-typedef void InAppStoreProductDelegate;
-typedef void InAppStoreTransactionObserver;
-#endif
-
-class InAppStore : public Object {
-	GDCLASS(InAppStore, Object);
-
-	static InAppStore *instance;
-	static void _bind_methods();
-
-	List<Variant> pending_events;
-
-	InAppStoreProductDelegate *products_request_delegate;
-	InAppStoreTransactionObserver *transactions_observer;
-
-public:
-	Error request_product_info(Dictionary p_params);
-	Error restore_purchases();
-	Error purchase(Dictionary p_params);
-
-	int get_pending_event_count();
-	Variant pop_pending_event();
-	void finish_transaction(String product_id);
-	void set_auto_finish_transaction(bool b);
-
-	void _post_event(Variant p_event);
-	void _record_purchase(String product_id);
-
-	static InAppStore *get_singleton();
-
-	InAppStore();
-	~InAppStore();
-};
-
-#endif
-
-#endif
+@end
