+++ modules/icloud/icloud.mm
@@ -28,21 +28,11 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifdef ICLOUD_ENABLED
-
 #include "icloud.h"
 
-#ifndef __IPHONE_9_0
-extern "C" {
-#endif
-
-#import "app_delegate.h"
+#import "platform/iphone/app_delegate.h"
 
 #import <Foundation/Foundation.h>
-
-#ifndef __IPHONE_9_0
-};
-#endif
 
 ICloud *ICloud::instance = NULL;
 
@@ -353,5 +343,3 @@ ICloud::ICloud() {
 }
 
 ICloud::~ICloud() {}
-
-#endif
