+++ modules/icloud/icloud_module.cpp
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  game_center.h                                                        */
+/*  icloud_module.cpp                                                    */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,48 +28,21 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifdef GAME_CENTER_ENABLED
+#include "icloud_module.h"
 
-#ifndef GAME_CENTER_H
-#define GAME_CENTER_H
+#include "core/config/engine.h"
 
-#include "core/object/class_db.h"
+#include "icloud.h"
 
-class GameCenter : public Object {
-	GDCLASS(GameCenter, Object);
+ICloud *icloud;
 
-	static GameCenter *instance;
-	static void _bind_methods();
+void register_icloud_types() {
+	icloud = memnew(ICloud);
+	Engine::get_singleton()->add_singleton(Engine::Singleton("ICloud", icloud));
+}
 
-	List<Variant> pending_events;
-
-	bool authenticated;
-
-	void return_connect_error(const char *p_error_description);
-
-public:
-	Error authenticate();
-	bool is_authenticated();
-
-	Error post_score(Dictionary p_score);
-	Error award_achievement(Dictionary p_params);
-	void reset_achievements();
-	void request_achievements();
-	void request_achievement_descriptions();
-	Error show_game_center(Dictionary p_params);
-	Error request_identity_verification_signature();
-
-	void game_center_closed();
-
-	int get_pending_event_count();
-	Variant pop_pending_event();
-
-	static GameCenter *get_singleton();
-
-	GameCenter();
-	~GameCenter();
-};
-
-#endif
-
-#endif
+void unregister_icloud_types() {
+	if (icloud) {
+		memdelete(icloud);
+	}
+}
