+++ modules/inappstore/in_app_store.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  view_controller.h                                                    */
+/*  in_app_store.h                                                       */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,18 +28,50 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#import <UIKit/UIKit.h>
+#ifndef IN_APP_STORE_H
+#define IN_APP_STORE_H
 
-@class GodotView;
-@class GodotNativeVideoView;
+#include "core/object/class_db.h"
 
-@interface ViewController : UIViewController
+#ifdef __OBJC__
+@class GodotProductsDelegate;
+@class GodotTransactionsObserver;
 
-@property(nonatomic, readonly, strong) GodotView *godotView;
-@property(nonatomic, readonly, strong) GodotNativeVideoView *videoView;
+typedef GodotProductsDelegate InAppStoreProductDelegate;
+typedef GodotTransactionsObserver InAppStoreTransactionObserver;
+#else
+typedef void InAppStoreProductDelegate;
+typedef void InAppStoreTransactionObserver;
+#endif
 
-// MARK: Native Video Player
+class InAppStore : public Object {
+	GDCLASS(InAppStore, Object);
 
-- (BOOL)playVideoAtPath:(NSString *)filePath volume:(float)videoVolume audio:(NSString *)audioTrack subtitle:(NSString *)subtitleTrack;
+	static InAppStore *instance;
+	static void _bind_methods();
 
-@end
+	List<Variant> pending_events;
+
+	InAppStoreProductDelegate *products_request_delegate;
+	InAppStoreTransactionObserver *transactions_observer;
+
+public:
+	Error request_product_info(Dictionary p_params);
+	Error restore_purchases();
+	Error purchase(Dictionary p_params);
+
+	int get_pending_event_count();
+	Variant pop_pending_event();
+	void finish_transaction(String product_id);
+	void set_auto_finish_transaction(bool b);
+
+	void _post_event(Variant p_event);
+	void _record_purchase(String product_id);
+
+	static InAppStore *get_singleton();
+
+	InAppStore();
+	~InAppStore();
+};
+
+#endif
