+++ platform/iphone/SCsub
@@ -8,9 +8,6 @@ iphone_lib = [
     "main.m",
     "app_delegate.mm",
     "view_controller.mm",
-    "game_center.mm",
-    "in_app_store.mm",
-    "icloud.mm",
     "ios.mm",
     "vulkan_context_iphone.mm",
     "display_server_iphone.mm",
