+++ platform/iphone/view_controller.mm
@@ -36,6 +36,7 @@
 #import "native_video_view.h"
 #include "os_iphone.h"
 
+#import <AVFoundation/AVFoundation.h>
 #import <GameController/GameController.h>
 
 @interface ViewController ()
@@ -213,15 +214,5 @@
 		return [self.videoView playVideoAtPath:filePath volume:videoVolume audio:audioTrack subtitle:subtitleTrack];
 	}
 }
-
-// MARK: Delegates
-
-#ifdef GAME_CENTER_ENABLED
-- (void)gameCenterViewControllerDidFinish:(GKGameCenterViewController *)gameCenterViewController {
-	//[gameCenterViewController dismissViewControllerAnimated:YES completion:^{GameCenter::get_singleton()->game_center_closed();}];//version for signaling when overlay is completely gone
-	GameCenter::get_singleton()->game_center_closed();
-	[gameCenterViewController dismissViewControllerAnimated:YES completion:nil];
-}
-#endif
 
 @end
